app/requirements.txt
flask
flask-cors
flask-jwt-extended
passlib[bcrypt]
pinecone
google-genai
dependency-injector
python-dotenv
numpy
onnxruntime
Pillow
